*** empty log message ***

+++ local/model/src/main/java/com/thinkparity/ophelia/model/profile/ProfileModelImpl.java
@@ -60,8 +60,7 @@ public final class ProfileModelImpl extends Model<ProfileListener> implements
     private static int isInviteUserEnabled;
 
     static {
-        DU_BACKUP_ALLOTMENT =
-            Long.valueOf(Integer.valueOf(1024 * 1024 * 1024 * 10).longValue());
+        DU_BACKUP_ALLOTMENT = Long.valueOf(1024 * 1024 * 1024 * 10);
     }
 
     /** A <code>ProfileEventGenerator</code> for local events. */
